# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/workflow/workflow/{id}": {
      "get": {
        "operationId": "WorkflowController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      },
      "put": {
        "operationId": "WorkflowController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsWorkflowDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      },
      "delete": {
        "operationId": "WorkflowController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsWorkflowDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow"
        ]
      }
    },
...
    "/auth/workflow/workflow_report/list": {
      "post": {
        "operationId": "WorkflowReportController_findAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseQueryDtoSmart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for list ToolsWorkflowReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolsWorkflowReportDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_report"
        ]
      }
    },
    "/auth/workflow/workflow_report/{id}": {
      "get": {
        "operationId": "WorkflowReportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowReportDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowReportDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_report"
        ]
      }
    },
...
    "/auth/workflow_aggregator/workflow_aggregator": {
      "get": {
        "operationId": "WorkflowAggregatorController_findAll",
        "parameters": [
          {
            "name": "pageIndex",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "JSON.stringify({})",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortKey",
            "required": false,
            "in": "query",
            "schema": {
              "example": "created_at",
              "type": "string"
            }
          },
          {
            "name": "sortValue",
            "required": false,
            "in": "query",
            "schema": {
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search_field",
            "required": false,
            "in": "query",
            "description": "Array of searchable fields",
            "schema": {
              "example": [
                "nama",
                "alamat"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "search_keyword",
            "required": false,
            "in": "query",
            "schema": {
              "example": "makmur",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for list ToolsWorkflowAggregatorDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 123
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_aggregator"
        ]
      },
      "post": {
        "operationId": "WorkflowAggregatorController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create ToolsWorkflowAggregatorDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_aggregator"
        ]
      }
    },
    "/auth/workflow_aggregator/workflow_aggregator/{id}": {
      "get": {
        "operationId": "WorkflowAggregatorController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for get ToolsWorkflowAggregatorDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_aggregator"
        ]
      },
      "put": {
        "operationId": "WorkflowAggregatorController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response for put ToolsWorkflowAggregatorDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_aggregator"
        ]
      },
      "delete": {
        "operationId": "WorkflowAggregatorController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response for delete ToolsWorkflowAggregatorDto",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "type": "number",
                      "example": 200
                    },
                    "pesan": {
                      "type": "string",
                      "example": "Berhasil"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ToolsWorkflowAggregatorDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflow_aggregator"
        ]
      }
    },
...
  },
  "info": {
    "title": "ERP API",
    "description": "Dokumentasi API ERP multi-tenant",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BaseQueryDtoSmart": {
        "type": "object",
        "properties": {
          "filter": {
            "type": "object",
            "description": "Where filter (supports >, >=, <, <=, like, notLike,  in, notIn, isNull, isNotNull)",
            "example": {
              "id": {
                "in": [
                  "id1",
                  "id2"
                ]
              },
              "status": {
                "eq": "aktif"
              },
              "created_at": {
                "between": [
                  "2024-01-01",
                  "2024-12-31"
                ]
              }
            },
            "additionalProperties": true
          },
          "include": {
            "type": "array",
            "description": "Nama tabel beserta tipenya single atau array",
            "example": [
              {
                "name": "contract",
                "type": "single",
                "select": [
                  "idContract",
                  "idClient",
                  "nomorKontrak",
                  "statusKontrak"
                ]
              },
              {
                "name": "client",
                "parent": "contract",
                "type": "single"
              },
              {
                "name": "work_schedule",
                "type": "array"
              }
            ],
            "additionalProperties": true
          },
          "joinWhere": {
            "type": "array",
            "description": "Filter + jenis JOIN (inner/left/right)",
            "example": [
              {
                "contract": {
                  "status_kontrak": "aktif"
                },
                "type": "inner"
              },
              {
                "client_site": {
                  "kota": "Jakarta"
                },
                "type": "left"
              }
            ]
          },
          "search_field": {
            "description": "Fields to search with search_keyword",
            "example": [
              "contract.nomor_kontrak",
              "site.kota"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "search_keyword": {
            "type": "string",
            "description": "Search keyword",
            "example": "PT"
          },
          "sortKey": {
            "type": "string",
            "description": "Sort by field",
            "example": "contract.nomor_kontrak"
          },
          "sortValue": {
            "type": "string",
            "description": "Sort direction (asc or desc)",
            "example": "asc"
          },
          "pageIndex": {
            "type": "number",
            "description": "Pagination page index",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "Pagination limit per page",
            "example": 20
          }
        }
      },
...
      "ToolsWorkflowDto": {
        "type": "object",
        "properties": {
          "idWorkflow": {
            "type": "string"
          },
          "namaWorkflow": {
            "type": "string"
          },
          "forModule": {
            "type": "string"
          },
          "targetDb": {
            "type": "string"
          },
          "targetTable": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "idKantor": {
            "type": "string"
          },
          "keterangan": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "statusWorkflow": {
            "type": "string"
          }
        },
        "required": [
          "idWorkflow",
          "namaWorkflow",
          "forModule",
          "targetDb",
          "targetTable"
        ]
      },
...
      "ToolsWorkflowReportDto": {
        "type": "object",
        "properties": {
          "idWorkflow": {
            "type": "string"
          },
          "namaWorkflow": {
            "type": "string"
          },
          "forModule": {
            "type": "string"
          },
          "targetDb": {
            "type": "string"
          },
          "targetTable": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "idKantor": {
            "type": "string"
          },
          "keterangan": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "statusWorkflow": {
            "type": "string"
          }
        },
        "required": [
          "idWorkflow",
          "namaWorkflow",
          "forModule",
          "targetDb",
          "targetTable"
        ]
      },
...
      "ToolsWorkflowAggregatorDto": {
        "type": "object",
        "properties": {
          "idWorkflowAggregator": {
            "type": "string"
          },
          "fromModule": {
            "type": "string"
          },
          "fromModuleId": {
            "type": "string"
          },
          "idWorkflow": {
            "type": "string"
          },
          "lastWorkflowStep": {
            "type": "string"
          },
          "lastApprovalNote": {
            "type": "string"
          },
          "lastApprovalUser": {
            "type": "string"
          },
          "lastApprovalDate": {
            "type": "string"
          },
          "lastApprovalStatus": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "idWorkflowAggregator",
          "fromModule",
          "fromModuleId",
          "idWorkflow"
        ]
      },
...
    }
  }
}